+++ bin/travis/doctrine_orm.sh
@@ -1,9 +1,8 @@
 #!/bin/bash
-
 DIR_NAME=`dirname $0`
 CONSOLE_DIR=$DIR_NAME"/.."
 
 # composer install --dev
-php $CONSOLE_DIR"/console" doctrine:database:create --env=orm
-php $CONSOLE_DIR"/console" doctrine:schema:create --env=orm
-php $CONSOLE_DIR"/console" doctrine:phpcr:repository:init
+$CONSOLE_DIR"/console" doctrine:database:create --env=orm
+$CONSOLE_DIR"/console" doctrine:schema:create --env=orm
+$CONSOLE_DIR"/console" doctrine:phpcr:repository:init