+++ bin/travis/phpcr_odm_doctrine_dbal.sh
@@ -1,8 +1,13 @@
 #!/bin/bash
+# make the script display the commands it runs to help debugging failures
+set -x
 
 DIR_NAME=`dirname $0`
 CONSOLE_DIR=$DIR_NAME"/.."
 
-# composer install --dev
-php $CONSOLE_DIR"/console" doctrine:phpcr:init:dbal --drop
-php $CONSOLE_DIR"/console" doctrine:phpcr:repository:init
+if ! $CONSOLE_DIR"/console" doctrine:phpcr:init:dbal --drop --force; then
+    # To support Jackalope <1.2
+    $CONSOLE_DIR"/console" doctrine:phpcr:init:dbal --drop
+fi
+
+$CONSOLE_DIR"/console" doctrine:phpcr:repository:init