+++ src/Functional/BaseTestCase.php
@@ -13,6 +13,7 @@
 namespace Symfony\Cmf\Component\Testing\Functional;
 
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\DependencyInjection\Container;
 
 /**
@@ -120,5 +121,22 @@ abstract class BaseTestCase extends WebTestCase
         }
 
         return parent::createKernel($options);
+    }
+
+    protected function assertResponseSuccess(Response $response)
+    {
+        libxml_use_internal_errors(true);
+
+        $dom = new \DomDocument();
+        $dom->loadHTML($response->getContent());
+
+        $xpath = new \DOMXpath($dom);
+        $result = $xpath->query('//div[contains(@class,"text-exception")]/h1');
+        $exception = null;
+        if ($result->length) {
+            $exception = $result->item(0)->nodeValue;
+        }
+
+        $this->assertEquals(200, $response->getStatusCode(), $exception ? 'Exception: "'.$exception.'"' : null);
     }
 }