+++ CHANGELOG.md
@@ -1,7 +1,30 @@
 Changelog
 =========
 
-* **2013-04-02**: [DEPRECATE] Deprecated `Content` document, it will be
-  removed in 2.0
+1.2.0-RC2
+---------
+
+* **2014-10-06**: Added purgeDatabase method to ORM manager
+
+1.2.0-RC1
+---------
+
+* **2014-07-27**: Added DataFixture support to the ORM DbManager
+* **2014-06-16**: Initializer for phpcr fixture loading
+* **2014-06-06**: Updated to PSR-4 autoloading
+
+1.1.0-RC2
+---------
+
+* **2014-06-11**: fetch manager by its name
+* **2014-04-24**: [DEPRECATE] Deprecated the `sonata_admin` bundle set
+* **2014-04-24**: added bundle sets: `sonata_admin_orm` and `sonata_admin_phpcr`
+* **2014-04-11**: dropped Symfony 2.2 compatibility
+
+1.1.0-RC1
+---------
+
+* **2013-04-07**: [DEPRECATE] Deprecated `LoadBaseData` DataFixture, it will be removed in 2.0
+* **2013-04-02**: [DEPRECATE] Deprecated `Content` document, it will be removed in 2.0
 * **2013-12-27**: Added `XmlSchemaTestCase` to test XML schema's
 * **2013-11-17**: Added `DatabaseTestListener` to support database testing